+++ tests/test_isort.py
@@ -1784,14 +1784,14 @@ def test_placement_control() -> None:
 
     assert test_output == (
         "import os\n"
+        "import p24.imports._argparse as argparse\n"
+        "import p24.imports._subprocess as subprocess\n"
         "import sys\n"
-        "from p24.imports import _VERSION as VERSION\n"
-        "from p24.imports import _argparse as argparse\n"
-        "from p24.imports import _subprocess as subprocess\n"
         "\n"
         "from bottle import Bottle, redirect, response, run\n"
         "\n"
-        "from p24.shared import media_wiki_syntax as syntax\n"
+        "import p24.imports._VERSION as VERSION\n"
+        "import p24.shared.media_wiki_syntax as syntax\n"
     )
 
 
@@ -1836,10 +1836,9 @@ def test_custom_sections() -> None:
     assert test_output == (
         "# Standard Library\n"
         "import os\n"
+        "import p24.imports._argparse as argparse\n"
+        "import p24.imports._subprocess as subprocess\n"
         "import sys\n"
-        "from p24.imports import _VERSION as VERSION\n"
-        "from p24.imports import _argparse as argparse\n"
-        "from p24.imports import _subprocess as subprocess\n"
         "\n"
         "# Django\n"
         "from django.conf import settings\n"
@@ -1853,7 +1852,8 @@ def test_custom_sections() -> None:
         "import pandas as pd\n"
         "\n"
         "# First Party\n"
-        "from p24.shared import media_wiki_syntax as syntax\n"
+        "import p24.imports._VERSION as VERSION\n"
+        "import p24.shared.media_wiki_syntax as syntax\n"
     )
 
 
@@ -4061,9 +4061,10 @@ def test_isort_ensures_blank_line_between_import_and_comment() -> None:
         "import one.b\n"
         "\n"
         "# noinspection PyUnresolvedReferences\n"
+        "import two.a as aa\n"
+        "\n"
         "# noinspection PyUnresolvedReferences\n"
-        "from two import a as aa\n"
-        "from two import b as bb\n"
+        "import two.b as bb\n"
         "\n"
         "# noinspection PyUnresolvedReferences\n"
         "from three.a import a\n"
@@ -4810,8 +4811,8 @@ def test_as_imports_mixed():
     test_input = """from datetime import datetime
 import datetime.datetime as dt
 """
-    expected_output = """from datetime import datetime
-from datetime import datetime as dt
+    expected_output = """import datetime.datetime as dt
+from datetime import datetime
 """
     assert isort.code(test_input, keep_direct_and_as_imports=True) == expected_output